+++ tests/Order/OrderServiceTest.php
@@ -61,37 +61,6 @@ final class OrderServiceTest extends ApiTestCase
         $this->assertCount(0, $secondItem->getDeclinationOptions());
     }
 
-    public function testCreateOrderReturn()
-    {
-        $orderService = $this->buildOrderService();
-
-        $creationCommand = new CreateOrderReturn(1, "Broken on arrival");
-        $creationCommand->addItem('1_0', 1, 1);
-
-        $returnId = $orderService->createOrderReturn($creationCommand);
-        $this->assertGreaterThan(0, $returnId);
-
-        $return = $orderService->getOrderReturn($returnId);
-        $this->assertSame($returnId, $return->getId());
-        $this->assertSame(1, $return->getOrderId());
-        $this->assertSame('R', $return->getStatus());
-        $this->assertSame('Broken on arrival', $return->getComments());
-        $this->assertGreaterThan(1500000000, $return->getCreatedAt()->getTimestamp());
-        $returnItems = $return->getItems();
-        $this->assertCount(1, $returnItems);
-        /** @var ReturnItem $returnItem */
-        $returnItem = reset($returnItems);
-        $this->assertSame(1, $returnItem->getAmount());
-        $this->assertSame('1_0', $returnItem->getDeclinationId());
-        $this->assertSame(67.9, $returnItem->getPrice());
-        $this->assertSame('Z11 Plus Boîtier PC en Acier ATX', $returnItem->getProductName());
-        $this->assertSame(1, $returnItem->getReason());
-
-        $returns = $orderService->getOrderReturns();
-        $this->assertCount(1, $returns);
-        $this->assertEquals($return, $returns[0]);
-    }
-
     public function testGetOrdersWithoutAuthentication()
     {
         $this->expectException(AuthenticationRequired::class);
@@ -102,41 +71,6 @@ final class OrderServiceTest extends ApiTestCase
     {
         $this->expectException(AuthenticationRequired::class);
         $this->buildOrderServiceWithoutAuthentication()->getOrder(1);
-    }
-
-    public function testGetOrderReturnWithoutAuthentication()
-    {
-        $this->expectException(AuthenticationRequired::class);
-        $this->buildOrderServiceWithoutAuthentication()->getOrderReturn(1);
-    }
-
-    public function testGetOrderReturnsWithoutData()
-    {
-        $returns = $this->buildOrderService()->getOrderReturns();
-
-        $this->assertCount(0, $returns);
-    }
-
-    public function testGetOrderReturnsWithoutAuthentication()
-    {
-        $this->expectException(AuthenticationRequired::class);
-        $this->buildOrderServiceWithoutAuthentication()->getOrderReturns();
-    }
-
-    public function testCreateOrderReturnWithoutAuthentication()
-    {
-        $this->expectException(AuthenticationRequired::class);
-        $this->buildOrderServiceWithoutAuthentication()->createOrderReturn(new CreateOrderReturn(1, ""));
-    }
-
-    public function testGetReturnReasons()
-    {
-        $returnTypes = $this->buildOrderServiceWithoutAuthentication()->getReturnReasons();
-        $this->assertCount(7, $returnTypes);
-
-        $this->assertSame(3, $returnTypes[0]->getId());
-        $this->assertSame('Colis arrivé en mauvais état, endommagé', $returnTypes[0]->getName());
-        $this->assertSame(10, $returnTypes[0]->getPosition());
     }
 
     public function testSendingAnAfterSalesServiceRequest()